Changed the error type that gets raised when a 503 status code is returned when polling for order status.
When polling for error status the method automatically tries again in 30
seconds if it doesn't get a 200 series return code.

With the new error type, the calling method will now catch the error and
try 5 times, at 30 second intervals.  The end result is the script will
now poll the serive for a full 10 minutes (30 second interval) before it
accepts failure.