Linked Parameters
When published parameters are linked, the python code that retrieves the
the values of a linked parameter gets just the name of the parameter to
which it is linked.

This revision has fixed that so that it will detect that the returned
value in a published parameter is in fact the name of antoher variable
and will work its way up the stack until it retrieves an actual value
and not a link.