Added code that will hopefully allow for PMP failovers when its not available.
PMP failover code existed however it was not implemented for every
method that interacts with PMP. That should now be resolved.  This
combined with fixing enhanced logging should help a great deal with if
not fixing PMP communication problems, then identifying what the
problems are with attempts to switch to failover.