Fixing problem with zipfile extraction in parcelmap transformer
Previous versions saw us move to extract from the packaged product that
LTSA is generating.  The packaged product zip file has a different
structure than the previous downloads. This impacts the ability of the
arcpy repolicator and the other dependent scritps from using what is
downloaded.  This commit contains the fix to that problem.  The new zip
method unzipPackagedProduct() will extract to a temp directory, then
move the gdb from the temp directory to the path that is expected by
other downstream processes