DWM reporter now has functionality to include failed features
Dwm reporter will check for the existence of an ffs file.  If it exists
then the failed features will be set to the number of features inside of
the ffs file.

Added new module FFSReader which includes the code required to read ffs
files.  May at a later point move this file to a library.